Split cluster into multiple files
